+++ test/Aqua.Tests/Serialization/TypeSystem/TypeInfo/When_using_typeinfo_with_circular_reference_no_propertyinfos.cs
@@ -9,6 +9,7 @@ using Xunit;
 
 public abstract class When_using_typeinfo_with_circular_reference_no_propertyinfos
 {
+#if !NET8_0_OR_GREATER
     public class With_binary_formatter : When_using_typeinfo_with_circular_reference_no_propertyinfos
     {
         public With_binary_formatter()
@@ -16,6 +17,8 @@ public abstract class When_using_typeinfo_with_circular_reference_no_propertyinf
         {
         }
     }
+
+#endif // NET8_0_OR_GREATER
 
     public class With_data_contract_serializer : When_using_typeinfo_with_circular_reference_no_propertyinfos
     {